this version has enhanced the profile screen, and separate component from screens'file.
